Changed the structure of the entire layout
